rename register_test -> registry_test & improve it!
- move all spec under top-level describe(),
  so that they can easily reuse the 'unregister' wrapper,
- fixup spec naming
